Refactor jdbc plugin to use MethodHandleDispatcher
