+++ apm-agent-plugins/apm-jdbc-plugin/src/main/java/co/elastic/apm/agent/jdbc/StatementInstrumentation.java
@@ -25,9 +25,8 @@
 package co.elastic.apm.agent.jdbc;
 
 import co.elastic.apm.agent.bci.VisibleForAdvice;
-import co.elastic.apm.agent.impl.ElasticApmTracer;
+import co.elastic.apm.agent.bootstrap.MethodHandleDispatcher;
 import co.elastic.apm.agent.impl.transaction.Span;
-import co.elastic.apm.agent.jdbc.helper.JdbcHelper;
 import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.NamedElement;
 import net.bytebuddy.description.method.MethodDescription;
@@ -55,8 +54,7 @@ public abstract class StatementInstrumentation extends JdbcInstrumentation {
 
     private final ElementMatcher<? super MethodDescription> methodMatcher;
 
-    StatementInstrumentation(ElasticApmTracer tracer, ElementMatcher<? super MethodDescription> methodMatcher) {
-        super(tracer);
+    StatementInstrumentation(ElementMatcher<? super MethodDescription> methodMatcher) {
         this.methodMatcher = methodMatcher;
     }
 
@@ -90,8 +88,8 @@ public abstract class StatementInstrumentation extends JdbcInstrumentation {
     @SuppressWarnings("DuplicatedCode")
     public static class ExecuteWithQueryInstrumentation extends StatementInstrumentation {
 
-        public ExecuteWithQueryInstrumentation(ElasticApmTracer tracer) {
-            super(tracer,
+        public ExecuteWithQueryInstrumentation() {
+            super(
                 named("execute").or(named("executeQuery"))
                     .and(takesArgument(0, String.class))
                     .and(isPublic())
@@ -101,48 +99,25 @@ public abstract class StatementInstrumentation extends JdbcInstrumentation {
         @Nullable
         @VisibleForAdvice
         @Advice.OnMethodEnter(suppress = Throwable.class)
-        public static Span onBeforeExecute(@Advice.This Statement statement,
-                                           @Advice.Argument(0) String sql) {
+        private static Span onBeforeExecute(@Advice.Origin Class<?> clazz,
+                                            @Advice.This Statement statement,
+                                            @Advice.Argument(0) String sql) throws Throwable {
 
-            if (tracer == null || jdbcHelperManager == null) {
-                return null;
-            }
-
-            JdbcHelper helperImpl = jdbcHelperManager.getForClassLoaderOfClass(Statement.class);
-            if (helperImpl == null) {
-                return null;
-            }
-
-            return helperImpl.createJdbcSpan(sql, statement, tracer.getActive(), false);
+            return (Span) MethodHandleDispatcher
+                .getMethodHandle(clazz, "co.elastic.apm.agent.jdbc.helper.AdviceHelperAdapter#createJdbcSpan")
+                .invoke(sql, statement, false);
         }
 
 
         @VisibleForAdvice
         @Advice.OnMethodExit(suppress = Throwable.class, onThrowable = Throwable.class)
-        public static void onAfterExecute(@Advice.This Statement statement,
-                                          @Advice.Enter @Nullable Span span,
-                                          @Advice.Thrown @Nullable Throwable t) {
-            if (span == null) {
-                return;
-            }
-
-            if (t == null && jdbcHelperManager != null) {
-                JdbcHelper helper = jdbcHelperManager.getForClassLoaderOfClass(Statement.class);
-                if (helper != null) {
-                    long count = helper.safeGetUpdateCount(statement);
-                    if (count != Long.MIN_VALUE) {
-                        span.getContext()
-                            .getDb()
-                            .withAffectedRowsCount(count);
-                    }
-                }
-
-            }
-
-            span.captureException(t)
-                .deactivate()
-                .end();
-
+        private static void onAfterExecute(@Advice.Origin Class<?> clazz,
+                                           @Advice.This Statement statement,
+                                           @Advice.Enter @Nullable Span span,
+                                           @Advice.Thrown @Nullable Throwable t) throws Throwable {
+            MethodHandleDispatcher
+                .getMethodHandle(clazz, "co.elastic.apm.agent.jdbc.helper.AdviceHelperAdapter#onAfterExecuteQuery")
+                .invoke(statement, span, t);
         }
     }
 
@@ -161,8 +136,8 @@ public abstract class StatementInstrumentation extends JdbcInstrumentation {
      */
     public static class ExecuteUpdateWithQueryInstrumentation extends StatementInstrumentation {
 
-        public ExecuteUpdateWithQueryInstrumentation(ElasticApmTracer tracer) {
-            super(tracer,
+        public ExecuteUpdateWithQueryInstrumentation() {
+            super(
                 named("executeUpdate").or(named("executeLargeUpdate"))
                     .and(takesArgument(0, String.class))
                     .and(isPublic())
@@ -172,39 +147,25 @@ public abstract class StatementInstrumentation extends JdbcInstrumentation {
         @Nullable
         @VisibleForAdvice
         @Advice.OnMethodEnter(suppress = Throwable.class)
-        public static Span onBeforeExecute(@Advice.This Statement statement,
-                                           @Advice.Argument(0) String sql) {
-            if (tracer == null || jdbcHelperManager == null) {
-                return null;
-            }
-
-            JdbcHelper helperImpl = jdbcHelperManager.getForClassLoaderOfClass(Statement.class);
-            if (helperImpl == null) {
-                return null;
-            }
-
-            return helperImpl.createJdbcSpan(sql, statement, tracer.getActive(), false);
+        private static Span onBeforeExecute(@Advice.Origin Class<?> clazz,
+                                            @Advice.This Statement statement,
+                                            @Advice.Argument(0) String sql) throws Throwable {
+            return (Span) MethodHandleDispatcher
+                .getMethodHandle(clazz, "co.elastic.apm.agent.jdbc.helper.AdviceHelperAdapter#createJdbcSpan")
+                .invoke(sql, statement, false);
         }
 
         @VisibleForAdvice
         @Advice.OnMethodExit(suppress = Throwable.class, onThrowable = Throwable.class)
-        public static void onAfterExecute(@Advice.Enter @Nullable Span span,
-                                          @Advice.Thrown @Nullable Throwable t,
-                                          @Advice.Return long returnValue /* bytebuddy converts int to long for us here ! */) {
-            if (span == null) {
-                return;
-            }
-
-            if (t == null) {
-                span.getContext()
-                    .getDb()
-                    .withAffectedRowsCount(returnValue);
-            }
-
-            span.captureException(t)
-                .deactivate()
-                .end();
+        private static void onAfterExecute(@Advice.Origin Class<?> clazz,
+                                           @Advice.Enter @Nullable Span span,
+                                           @Advice.Thrown @Nullable Throwable t,
+                                           @Advice.Return long returnValue /* bytebuddy converts int to long for us here ! */) throws Throwable {
+            MethodHandleDispatcher
+                .getMethodHandle(clazz, "co.elastic.apm.agent.jdbc.helper.AdviceHelperAdapter#onAfter")
+                .invoke(span, t, returnValue);
         }
+
     }
 
     /**
@@ -212,8 +173,8 @@ public abstract class StatementInstrumentation extends JdbcInstrumentation {
      */
     public static class AddBatchInstrumentation extends StatementInstrumentation {
 
-        public AddBatchInstrumentation(ElasticApmTracer tracer) {
-            super(tracer,
+        public AddBatchInstrumentation() {
+            super(
                 nameStartsWith("addBatch")
                     .and(takesArgument(0, String.class))
                     .and(isPublic())
@@ -221,15 +182,10 @@ public abstract class StatementInstrumentation extends JdbcInstrumentation {
         }
 
         @Advice.OnMethodEnter(suppress = Throwable.class)
-        public static void storeSql(@Advice.This Statement statement, @Advice.Argument(0) String sql) {
-            if (jdbcHelperManager == null) {
-                return;
-            }
-
-            JdbcHelper helperImpl = jdbcHelperManager.getForClassLoaderOfClass(Statement.class);
-            if (helperImpl != null) {
-                helperImpl.mapStatementToSql(statement, sql);
-            }
+        private static void storeSql(@Advice.Origin Class<?> clazz, @Advice.This Statement statement, @Advice.Argument(0) String sql) throws Throwable {
+            MethodHandleDispatcher
+                .getMethodHandle(clazz, "co.elastic.apm.agent.jdbc.helper.AdviceHelperAdapter#mapStatementToSql")
+                .invoke(statement, sql);
         }
     }
 
@@ -237,14 +193,14 @@ public abstract class StatementInstrumentation extends JdbcInstrumentation {
 
     /**
      * Instruments:
-     *  <ul>
-     *      <li>{@link Statement#executeBatch()} </li>
-     *      <li>{@link Statement#executeLargeBatch()} (java8)</li>
-     *  </ul>
+     * <ul>
+     *     <li>{@link Statement#executeBatch()} </li>
+     *     <li>{@link Statement#executeLargeBatch()} (java8)</li>
+     * </ul>
      */
     public static class ExecuteBatchInstrumentation extends StatementInstrumentation {
-        public ExecuteBatchInstrumentation(ElasticApmTracer tracer) {
-            super(tracer,
+        public ExecuteBatchInstrumentation() {
+            super(
                 named("executeBatch").or(named("executeLargeBatch"))
                     .and(takesArguments(0))
                     .and(isPublic())
@@ -254,66 +210,33 @@ public abstract class StatementInstrumentation extends JdbcInstrumentation {
 
         @Nullable
         @Advice.OnMethodEnter(suppress = Throwable.class)
-        @SuppressWarnings("DuplicatedCode")
-        public static Span onBeforeExecute(@Advice.This Statement statement) {
-            if (tracer == null || jdbcHelperManager == null) {
-                return null;
-            }
-            JdbcHelper helper = jdbcHelperManager.getForClassLoaderOfClass(Statement.class);
-            if (helper == null) {
-                return null;
-            }
-
-            String sql = helper.retrieveSqlForStatement(statement);
-            return helper.createJdbcSpan(sql, statement, tracer.getActive(), true);
-
+        private static Span onBeforeExecute(@Advice.Origin Class<?> clazz, @Advice.This Statement statement) throws Throwable {
+            return (Span) MethodHandleDispatcher
+                .getMethodHandle(clazz, "co.elastic.apm.agent.jdbc.helper.AdviceHelperAdapter#createJdbcSpanLookupSql")
+                .invoke(statement, false);
         }
 
         @Advice.OnMethodExit(suppress = Throwable.class, onThrowable = Throwable.class)
-        public static void onAfterExecute(@Advice.Enter @Nullable Span span,
-                                          @Advice.Thrown Throwable t,
-                                          @Advice.Return Object returnValue) {
-            if (span == null) {
-                return;
-            }
-
-            // for 'executeBatch' and 'executeLargeBatch', we have to compute the sum as Statement.getUpdateCount()
-            // does not seem to return the sum of all elements. As we can use instanceof to check return type
-            // we do not need to use a separate advice. 'execute' return value is auto-boxed into a Boolean,
-            // but there is no extra allocation.
-            long affectedCount = 0;
-            if (returnValue instanceof int[]) {
-                int[] array = (int[]) returnValue;
-                for (int i = 0; i < array.length; i++) {
-                    affectedCount += array[i];
-                }
-            } else if (returnValue instanceof long[]) {
-                long[] array = (long[]) returnValue;
-                for (int i = 0; i < array.length; i++) {
-                    affectedCount += array[i];
-                }
-            }
-            span.getContext()
-                .getDb()
-                .withAffectedRowsCount(affectedCount);
-
-            span.captureException(t)
-                .deactivate()
-                .end();
+        private static void onAfterExecute(@Advice.Origin Class<?> clazz,
+                                           @Advice.Enter @Nullable Span span,
+                                           @Advice.Thrown Throwable t,
+                                           @Advice.Return Object returnValue) throws Throwable {
+            MethodHandleDispatcher
+                .getMethodHandle(clazz, "co.elastic.apm.agent.jdbc.helper.AdviceHelperAdapter#onAfterExecuteBatch")
+                .invoke(span, t, returnValue);
         }
-
     }
 
     /**
      * Instruments:
-     *  <ul>
-     *      <li>{@link PreparedStatement#executeUpdate()} </li>
-     *      <li>{@link PreparedStatement#executeLargeUpdate()} ()} (java8)</li>
-     *  </ul>
+     * <ul>
+     *     <li>{@link PreparedStatement#executeUpdate()} </li>
+     *     <li>{@link PreparedStatement#executeLargeUpdate()} ()} (java8)</li>
+     * </ul>
      */
     public static class ExecuteUpdateNoQueryInstrumentation extends StatementInstrumentation {
-        public ExecuteUpdateNoQueryInstrumentation(ElasticApmTracer tracer) {
-            super(tracer,
+        public ExecuteUpdateNoQueryInstrumentation() {
+            super(
                 named("executeUpdate").or(named("executeLargeUpdate"))
                     .and(takesArguments(0))
                     .and(isPublic())
@@ -322,39 +245,22 @@ public abstract class StatementInstrumentation extends JdbcInstrumentation {
 
         @Nullable
         @Advice.OnMethodEnter(suppress = Throwable.class)
-        @SuppressWarnings("DuplicatedCode")
-        public static Span onBeforeExecute(@Advice.This Statement statement) {
-            if (tracer == null || jdbcHelperManager == null) {
-                return null;
-            }
-
-            JdbcHelper helperImpl = jdbcHelperManager.getForClassLoaderOfClass(Statement.class);
-            if (helperImpl == null) {
-                return null;
-            }
-
-            String sql = helperImpl.retrieveSqlForStatement(statement);
-            return helperImpl.createJdbcSpan(sql, statement, tracer.getActive(), true);
+        private static Span onBeforeExecute(@Advice.Origin Class<?> clazz, @Advice.This PreparedStatement statement) throws Throwable {
+            return (Span) MethodHandleDispatcher
+                .getMethodHandle(clazz, "co.elastic.apm.agent.jdbc.helper.AdviceHelperAdapter#createJdbcSpanLookupSql")
+                .invoke(statement, true);
         }
 
         @Advice.OnMethodExit(suppress = Throwable.class, onThrowable = Throwable.class)
-        public static void onAfterExecute(@Advice.Enter @Nullable Span span,
-                                          @Advice.Thrown @Nullable Throwable t,
-                                          @Advice.Return long returnValue /* bytebuddy converts int to long for us here ! */) {
-            if (span == null) {
-                return;
-            }
-
-            if (t == null) {
-                span.getContext()
-                    .getDb()
-                    .withAffectedRowsCount(returnValue);
-            }
-
-            span.captureException(t)
-                .deactivate()
-                .end();
+        private static void onAfterExecute(@Advice.Origin Class<?> clazz,
+                                           @Advice.Enter @Nullable Span span,
+                                           @Advice.Thrown @Nullable Throwable t,
+                                           @Advice.Return long returnValue /* bytebuddy converts int to long for us here ! */) throws Throwable {
+            MethodHandleDispatcher
+                .getMethodHandle(clazz, "co.elastic.apm.agent.jdbc.helper.AdviceHelperAdapter#onAfter")
+                .invoke(span, t, returnValue);
         }
+
     }
 
     /**
@@ -365,8 +271,8 @@ public abstract class StatementInstrumentation extends JdbcInstrumentation {
      * </ul>
      */
     public static class ExecutePreparedStatementInstrumentation extends StatementInstrumentation {
-        public ExecutePreparedStatementInstrumentation(ElasticApmTracer tracer) {
-            super(tracer,
+        public ExecutePreparedStatementInstrumentation() {
+            super(
                 named("execute").or(named("executeQuery"))
                     .and(takesArguments(0))
                     .and(isPublic())
@@ -375,42 +281,21 @@ public abstract class StatementInstrumentation extends JdbcInstrumentation {
 
         @Nullable
         @Advice.OnMethodEnter(suppress = Throwable.class)
-        @SuppressWarnings("DuplicatedCode")
-        public static Span onBeforeExecute(@Advice.This Statement statement) {
-            if (tracer != null && jdbcHelperManager != null) {
-                JdbcHelper helperImpl = jdbcHelperManager.getForClassLoaderOfClass(Statement.class);
-                if (helperImpl != null) {
-                    @Nullable String sql = helperImpl.retrieveSqlForStatement(statement);
-                    return helperImpl.createJdbcSpan(sql, statement, tracer.getActive(), true);
-                }
-            }
-            return null;
+        private static Span onBeforeExecute(@Advice.Origin Class<?> clazz, @Advice.This PreparedStatement statement) throws Throwable {
+            return (Span) MethodHandleDispatcher
+                .getMethodHandle(clazz, "co.elastic.apm.agent.jdbc.helper.AdviceHelperAdapter#createJdbcSpanLookupSql")
+                .invoke(statement, true);
         }
 
         @Advice.OnMethodExit(suppress = Throwable.class, onThrowable = Throwable.class)
-        public static void onAfterExecute(@Advice.This Statement statement,
-                                          @Advice.Enter @Nullable Span span,
-                                          @Advice.Thrown @Nullable Throwable t) {
-
-            if (span == null) {
-                return;
-            }
-
-            if (t == null && jdbcHelperManager != null) {
-                JdbcHelper jdbcHelper = jdbcHelperManager.getForClassLoaderOfClass(Statement.class);
-                if (jdbcHelper != null) {
-                    span.getContext()
-                        .getDb()
-                        // getUpdateCount javadoc indicates that this method should be called only once
-                        // however in practice adding this extra call seem to not have noticeable side effects
-                        .withAffectedRowsCount(jdbcHelper.safeGetUpdateCount(statement));
-                }
-            }
-
-            span.captureException(t)
-                .deactivate()
-                .end();
+        private static void onAfterExecute(@Advice.Origin Class<?> clazz,
+                                           @Advice.This Statement statement,
+                                           @Advice.Enter @Nullable Span span,
+                                           @Advice.Thrown @Nullable Throwable t) throws Throwable {
+            MethodHandleDispatcher
+                .getMethodHandle(clazz, "co.elastic.apm.agent.jdbc.helper.AdviceHelperAdapter#onAfterPreparedStatementExecuteQuery")
+                .invoke(statement, span, t);
         }
-
     }
+
 }
