+++ apm-agent-plugins/apm-jdbc-plugin/src/test/java/co/elastic/apm/agent/jdbc/AbstractJdbcInstrumentationTest.java
@@ -25,11 +25,11 @@
 package co.elastic.apm.agent.jdbc;
 
 import co.elastic.apm.agent.AbstractInstrumentationTest;
+import co.elastic.apm.agent.bootstrap.MethodHandleDispatcher;
 import co.elastic.apm.agent.impl.context.Db;
 import co.elastic.apm.agent.impl.context.Destination;
 import co.elastic.apm.agent.impl.transaction.Span;
 import co.elastic.apm.agent.impl.transaction.Transaction;
-import co.elastic.apm.agent.jdbc.helper.JdbcHelper;
 import co.elastic.apm.agent.jdbc.signature.SignatureParser;
 import org.junit.After;
 import org.junit.Before;
@@ -161,11 +161,12 @@ public abstract class AbstractJdbcInstrumentationTest extends AbstractInstrument
 
             // clear internal jdbc helper required due to metadata caching and global state about unsupported
             // JDBC driver features (based on classes instances)
-            if (JdbcInstrumentation.jdbcHelperManager != null) {
-                JdbcHelper jdbcHelper = JdbcInstrumentation.jdbcHelperManager.getForClassLoaderOfClass(Statement.class);
-                if (jdbcHelper != null) {
-                    jdbcHelper.clearInternalStorage();
-                }
+            try {
+                MethodHandleDispatcher
+                    .getMethodHandle(getClass(), "co.elastic.apm.agent.jdbc.helper.AdviceHelperAdapter#clearInternalStorage")
+                    .invoke();
+            } catch (Throwable throwable) {
+                throwable.printStackTrace();
             }
         }
     }
